// asset-management/src/generated/models/Cr240_assetmaintenancesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr240_assetmaintenancescr240_maintenancetype = {
  805530000: 'Preventive',
  805530001: 'Corrective',
  805530002: 'Predictive'
} as const;
export type Cr240_assetmaintenancescr240_maintenancetype = keyof typeof Cr240_assetmaintenancescr240_maintenancetype;
export const Cr240_assetmaintenancesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr240_assetmaintenancesstatecode = keyof typeof Cr240_assetmaintenancesstatecode;
export const Cr240_assetmaintenancesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr240_assetmaintenancesstatuscode = keyof typeof Cr240_assetmaintenancesstatuscode;

export interface Cr240_assetmaintenancesBase {
  "cr240_AssetID@odata.bind"?: string;
  cr240_assetmaintenanceid: string;
  cr240_cost?: number;
  cr240_maintenancedate?: string;
  cr240_maintenancename?: string;
  cr240_maintenancetype?: Cr240_assetmaintenancescr240_maintenancetype;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cr240_assetmaintenancesstatecode;
  statuscode?: Cr240_assetmaintenancesstatuscode;
  timezoneruleversionnumber?: number;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: number;
}

export interface Cr240_assetmaintenances extends Cr240_assetmaintenancesBase {
  cr240_assetidname?: string;
  cr240_cost_base?: number;
  cr240_maintenancetypename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  exchangerate?: number;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  transactioncurrencyidname?: string;
  versionnumber?: number;
  cr240_assetid?: object;
  _cr240_assetid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}